Start the simplification of the display parameter definition
For now the definition of the parameters used to define the type
and the sizes of the objects displayed in VTK is quite cumbersome.
The idea is to begin grouping the related parameters (units, camera,
and so on...) and then define ways to stablish those related parameters
toghether (i.e. use meters and kilonewtons to display all quantities)
without losing the possibility of a more fine-grained treatement.
That commit is only the first step: grouping the units related parameters
into a base class (so, for the moment, all works in the same way).