Calculate roughly a hyperbolic paraboloid shell (umbrella)
The class hyparUmbrella is added to vaults_rough_calc module. It performs
the rough calculation of a hypar shell according to method proposed in
edX course "The Art of Structural Engineering: Vaults", by Maria Garlock.
A test is added to check this class, taking as example the Rios Warehouse in
Mexico City, by Félix Candela.