Refactor graphics routines in order to simplify them
The idea is to allow to display graphics in a simple way as soon
as there is something to display. To achieve that we'll add the
appropriate methods and parameters to the modelSpace object. For
the moment I've added the methods that display the model geometry.
The next step is to put into the modelSpace object the parameters
needed to display forces, internal forces, reactions and so on.