Fix errors displaying load diagrams
The beam and shell elements have different labels to represent
load components so the routine works only when you use a label
that is valid for both kind of elements. Moreover the routine that
computes the maximum value of the load did not filter the elements
that belong to the set to display so, even if you select a set
without shell elements if the load to represent is also defined
on shell elements the program crashed. Now both types of elements
return a Vector3d object with the load components in local coordinates
which is then transformed to global coordinates using the unit vectors
of the element reference system.