Move compressive strength calculation methods to ASTMShape class
We need to have some way of calculating strenghts of the section
itself (without taking into account buckling effect) as we already
do with getReferenceFlexuralStrength. This way we can compute the
effects of buckling with the chiN and chiLT factors passed to
getBiaxialBendingMethod. The design intent is to mimic the Eurocode
method of separating the effects of the local buckling of the steel
shape section and the buckling of the member.