Add end grain case to lateral strength calculation
Where dowel-type fasteners with D > 1/4" are inserted into the
end grain of the main member, with the fastener axis parallel to
the wood fibers, F_perp shall be used in the determination of the dowel
bearing strength of the main member, F_em. See clause 12.3.3.4
of National Design Specification 2018 (American wood council).