Implementation of limit state checking for RC circular sections
Move the methods that create the sections to the base class. This
allow us to derive section shape agnostic classes from this base
class (i. e. circular shaped sections). On the other hand I've
renamed the creaTwoSections method as createSections: this allows
to use the method name to create any number of
sections: len(gaussPoints)*len(directions). Finally I've moved
the call to createSections into the append method of SectionContainer
because both method are allways called with this sequence.