Fix error with set boolean operators
The operators defined by Boost.Python return COPIES of the
object stored in MapSet. As a consequence the operations in
those objects are lost (they are simply temporary objects
dettached from its counterpart in MapSet). The solution
has been to remove the Boost.Python "built-in" operators
(+,- and *) and use only only their counterparts (+=, -=
and *=). I've also created two methods in predefined_spaces
to make things easier: setSum and setIntersect.