Add explanation about error in linear buckling analysis
The error comes from the element formulation in OpenSees to
capture the geometric non-linearity "inside" the element you need to
use a different element formulation, e.g., nonlinear strain in the
displacement-based formulation or curvature-based displacement
interpolation (CBDI) in the force-based formulation. Nevertheless
the results with a finer discretization are good enough.