Update test values
After fixing the problem with the 3D polygon constructor from
a point list the mesh of this tests have changed. The element
size on test paving 07 was not valid for the algorithm anymore
so I've changed it. This algorithm will probably be deprecated
because its very instable. For now the test work again.