Solve problem with definition in control vars in results files
Now we insert a string at the beginning of the results file
that import the module where the control var class is defined.
This way there is no necessity to import the appropriate module
before loading the results. I've also fixed a minor error concerning
the duplicate definition of 'chiLT' on the results file data.