Rename methods related to with-to-thickness ratios
There are some new methods to add from table D1.1 of AISC 341 and
the current methods names are vague. I've renamed them to clarify
it they correspond to compression ratios (table B4.1a of AISC 360-16)
or to bending ratios (table B4.1a of AISC 360-16). The new methods
implementing table D1.1 will be called LambdaHD and LambdaMD
corresponding to the LambdaP and LambdaR methods of tables
B4.1a and B4.1b of AISC 360.