Import base modules inside boost.python
This way there is no need that the user must know that to import
the "xc" module he needs to import the "geom" module first and
that to import this one he needs to import the "xc_base" module
first again. From now on this process is transparent to the
user.