Fix error in Segment3d PtoParametricas method
The method interpreted the lambda parameter as some sort of natural
coordinate (with values between 0 and 1) and was inconsistent with
the rest of the PtoParametricas method in the library. The error
was "compensated" in some of the Python modules which have been
fixed too (load_model_base.py, model/geometry/geom_utils.py and
section_by_folding_planes.py).