Don't try to check elements which has no defined section
Avoid calling checking routines for the elements which has no
defined reinforced concrete section. Moreover, instead of issue
a warning for each of those elements, group their identifiers
in a set and issue a single warning message at the end of the
routine.

The idea is to print only an error message with the identifiers
of all the elements with missing section names. This avoids to
encomber the error console with a long list of error messages.