Expose Bnd and In methods
Expose the methods that return the boundary box and the method
that check if a point is inside the boundary box to Python.
The idea was to use this in the embankment python module, but
finally this is not needed. Instead of that we check that the
intersection of a vertical line with the embankemnt layer exists.