Fix error in shape functions for quadrilateral surfaces
The expression used previously was valid only for rectangles
with its sides parallels to the coordinate axes. I have adapted
the getNaturalCoordinates method used in shell elements to compute
the natural coordinates of a given point and then the shape functions
are computed using the classic formulas expressed in natural
coordinates.