Implement offset and getBufferPolygon methods in PlanePolyline3d
Those methods are useful to obtain a contour from a 3D plane polyline
and use this contour to select nodes or elements. This technique
is used to compute the nodes or elements that are loaded by a
linear load which spreads to some soil layers (see load_model_base
module in railway_traffic package.