Implement clipping of the uniform rail loads (work in progress)
When the axis of the uniform rail load extends to both sides of
the loaded wall, th part of the rail behind the wall, creates
"negative" pressures. This is wrong (the backfill occupies only
the loaded side of the wall) so the solution is to clip the axis
of the rail with the wall mid-plane and remove the shortest of
the resultant polylines (we suppose that the longest part is the
one that the user wants to be loaded).