Fix error removing freeze constraints from nodes
Freeze constraints are created to freeze the nodes connected only
to dead elements (to avoid singularities in the system). When a
node has been frozen twice (because it is connected to two sets
of frozen elements) and a melt of the nodes is requested during
the reactivation of one of the sets we must take care of the nodes
which belong to both sets removing the constraint of those that
are connected to one or more «resurrected» elements. See for
example: kill_elements_09.py