Implement method to integrate the pressures along a surface
The method computes the sliding vectors system which is equivalent
to the pressures on the surface. This sliding vector system can be
used to load a series of nodes that doesn't need to be on the
surface used to to compute the pressures. In retaining walls, this
allows to compute the forces on a virtual back and then transfer
this loads to the nodes of the wall stem.