Enhance computation of permanent loads
Make the methods that compute the permenent loads on the wall to
compute the resulting sliding vector system so it can be used
to check the computed values. Introduce two length correction
factors in createDeadLoad that takes into account the differences
between the true lengths of the toe and the heel and their lengths
in the wireframe model.