Adapt the code to the requirements of Eurocode 7 calculations
Add the ignoreAdhesion parameter to control whether or no the adhesion
of cohesive soils to the concrete of the foundation must be considered
 when computing the sliding resistance of the foundation.

Allow using different partial factors for the bearing resistance and
for the sliding resistance.

Add method to return the degree of utilization obtained in the verification
of resistance to sliding.

Add smoothPrecastFoundation member which when true, means that a smooth
contact between the foundation and the soil.