Minor adjustement of the results due to an error fixed
The bottom end of the stem in the wireframe model was in the wrong
position because we added the whole width increment due to the stem
back slope. The correct value is obtained using half of this increment
because the point must be in the center of the bottom section.