Implement stepped stems in retaining walls and fix error
The error fixed was in the position of the bottom en of the stem.
This point was in the wrong position because we added the whole width
increment due to the stem back slope. The correct value is obtained
using half of this increment because the point must be in the center
of the bottom section.