Add information about the member type
The member type is used by the methods that compute the minimal
reinforcement of the section.

Moreover I've removed the limit in the bending resistance returned
by getMR because, even if this moment is a rough measure of the
bending resistance provided by the excess or deficiency of reinforcement
it's useful as a measure of the efficiency of the section reinforcement
area. If we mask this value with a limit the results become confusing
because the lack of sensibility to the variations in the reinforcement
area.