Add jointSpacing attribute and fix some errors
The jointSpacing attribute allows to check for the distance between
joints in the functions that return the minimal quantities of
reinforcement due to concrete shrinking.

The method getSection has been renamed as getSectionDepth to make
clear that the value returned corresponds to the thickness of the
stem a the "y" height. Moreover this method returned the wrong
value when the backface of the stem has two or more steps. This
error has been fixed.

The method getYStem has been moved here from the derived class
RetainingWall.

There was a bug in the getStemContourPoints method that returned
the wrong contour when the number of steps in the back face of the
stem was greater than one. The depth of each step was interpreted as
a cumulated value (not an incremental one) which was wrong. The name
of the variable that stores the increment has been changed from depth
to depthIncrement to make that clear.