Modify the method getMinReinfAreaInTension
Previously the method returned the value obtained by applying the
expression (7.1) of the clause 7.3.2 of EC2:2004 part 1. This has no
sense because there are other clauses that specify lower values
for this reinforcement area depending on the type of the member.
It seems logical to check the value specified for the particular
member type (work in progress) and if it is not found to fall back
into the expression (7.1) cited quoted above.