Consider the axial force in the footing in bending and shear check
Consider the favourable effect of the axial force when computing the
bending and shear resistances.

Add a method called getStemMemberType that returns 'wall' or
'short_wall' depending on the joint spacing. Add the member type
to the sections returned by the getSection* methods.

Remove prints used for debugging.

Interpolate the value of the maximum shear force in the wall footing
by assuming that the shear force decreases linearly to reach zero at
the end of the heel.

Return the minimum compression along the heel in getInternalForces.
Use the values obtained at both extremities of the element to compute
the values in the worst case.