Store the information returned by the solver in the "info" property
Some of the underlying FORTRAN or C solvers return an integer error
code that can help to identify the nature of the problem with the
stiffness matrix. This error code is stored in a property called
"info" which is reachable by Python. You can see the
locate_equation_number_01.py example to see how it works.