Fix error when importing model data from JSON
JSON uses strings as dictionary keys so, when the dictionary is
read from this format we need to convert the strings into integers.
The new function tag_integer_from_py_object in python_utils.h admits
a boost::python::str or boost::python::int object as arguments and
returns an integer.