Fix error in simplify method
Erasing items in the list still invalidates the following iterators
in the "selected" list (it works ok when there is only one item
to remove). The new code creates a new point sequence by ignoring
the items to remove and then assigns it to the calling object.
See remove_selected method.