Allow buffers with "flat" extremities
In some cases, for example, when creating loads around a polyline,
the buffer semicircles at the extremities are a source of problems.
Most of the time, the load needs to be finished right at the extremities
of the polyline. Those buffers around the end points make the load go
beyond those limits.