Implement checkTributaries method
This method checks if the internal nodes of the set have already
a tributary length (or area or volume) and, in that case, prevents
the user about this circumstance. The tributary attribute of the
nodes is cumulative (to avoid discontinuities in the nodes that
lie in the boudaries of a set), this can lead to errors if the
method is called twice on an interior node.