Fix error in topology of imported elements
The topology of the points, edges and surfaces was ignored because
the edges doesn't exist yet when the point was created, the surfaces
doesn't exist yet when the edge was created and so on.

The solution adopted is to ignore the fields storing the topology
and recreate it when the edges, surfaces and bodies are readed.