Add defDiag method for concrete materials
This method can define the XC material representing the stress-strain
diagram of the reinforcement steel with either characteristic or design
values. This makes the code consistent with other methods that receive
the diagram type ('k' or 'd') as parameter.

The defDiag method for reinforcing steel has been fixed. There was
a bug when calling the defDiagK and defDiagD methods.