Add convennience methods in PredefinedSpace
The methods newElement and newSeedElement simplify the creation of a
new element without needing to specify the element handler to use. There
is no need to specify the node tags for seed elements. In a future this
will issue a warning message to the user explaining that it's no use
to specify the node tags for seed elements.

The methods setDefaultMaterial and setDefaultCoordTransf simplify the
definition of those parameters with no need to specify the element handler
to use.