Fix error in create_nodes
The error occurs when, for whatever reason, the specified number
of divisions for the J sides of the quadrilateral (NDivJ) is different
than the actual number of divisions of the J sides objects (for example
when those sides have been meshed independently of the quadrilateral
face). The adopted solution is to check for that the number of divisions
of the sides and the quadrilateral face are consistent and, if not,
issue and error messages and adopt the value obtained from the sides.
Otherwise the program crashes due because the addresses obtained for
the nodes are incorrect.