Allow changing the default material for the seed element any time
Until now, the material needed to be specified before the seed element
was created, and the seed element material remained unchanged even if
the default material was no longer the same. From now, on the
specification of a new material is allowed at any moment and the
seed element is recreated as needed. In any case the material must be
compatible with the element type.