Fix wrong values and simplify code
The previously computed values were wrong because the solution procedure
used with the crack controller was linear static and the fiber sections
created in the phantom model are non-linear. This has been solved by adding
a parameter to the initializer of the controllers to specify the type
of solution procedure to use.