No craking when concrete strain too small
The previous code in this case was wrong. If the concrete stress is
under its tensile strenght there is no craking. Nevertheless, the
stress-strain model of the concrete in the fiber-section model has
no tensile strenght so the results when the bending moments are
too small the results are quite conservative.