Store the maximum and minimum stresses
Store the maximum stress in the reinforcement and the minimum stress
in the concrete. The maximum stress in the reinforcement can be used
as an indirect measurement of the crack width and the minimum stress
in the concrete can be used to check the limitations for this
parameter in some codes (serviceability limit state).