Allow the user define the wind orientation
The wind load will always be normal to the track axis, and it can
be reversed by changing the sign of the wind pressure. Nevertheless,
for thoshe axes whose curvature sign changes along their length,
you need to inform the load generator of the direction of the wind load.
This is done through the windDirection parameter (see
test_track_axis_load_07.py example).