Use non-linear solution algorithm for shear checking
The shear checking of reinforced concrete sections is computed using
the tension state of the fiber section under the corresponding internal
forces. To compute this tension state, it is needed a non-linear solution
algorithm because the concrete and steel materials are non-linear, so I've
changed the default solution algorithm accordingly.

The reference values of some test have changed due to this modification. The
required changes have been small except for the tests:
   ehe08/test_shear_uls_checking_05.py
   ehe08/test_shear_uls_checking_06.py
   ec2/test_shear_uls_checking_05.py
   ec2/test_shear_uls_checking_06.py
In those test the changes in the reference values due to convergence problems
with the previous value of the uniform load on the slab which, apparently,
was too big to achieve equilibrium in the RC section.