Adapt the reference values to the new solution algorithm
The shear checking of reinforced concrete sections is computed using
the tension state of the fiber section under the corresponding internal
forces. To compute this tension state, it is needed a non-linear solution
algorithm because the concrete and steel materials are non-linear, so I've
changed the default solution algorithm accordingly.

The reference values of some test have changed due to this modification.