Issue an error message in dot product
Issue an error message when the vectors in the dot product have not
the same size. Return the dot product using the smalles of the two
sizes. This method can be called from Python so is not a good idea
to issue the error message only for debugging versions of the code.