Move the intForcCombFileName parameter up in the execution thread
The idea is to read the internal forces data out of the phantom model
and build the phantom model from the tuple containing the element tags,
the identifiers of the load combinations and the values of the internal
forces. This way we can do some preprocessing of the internal forces
if needed. For example for buckling limit states we can increment the
internal forces to take into account the results of the buckling analysis
performed.