Issue a warning message when providing element node tags when not needed
Seed elements don't need node tags because they are used only as
a template element to put in a mesh using mesh generation commands.
To avoid confusion using seed element handler and regular element handler,
from now on a warning message is isued when the node tags are specified
for a seed element.