Use a more accurate expression for shear strength in compression
The value obtained with the new expression is very close to the
value obtained with a very small bending moment (see
test_fiber_section_shear_04.py).

This case is quite rare (compressed section without bending moment)
so the EHE standard does not consider it. The adopted solution seems
to be conservative enough.