Allow many situations to check for a limit state
Until now, by default, the ULS checking was performed only for
permanent and transient situations. The user was required to check
the earthquake and accidental combinations in separate analysis
using, perhaps, a different partial safety factors for materials.
This can lead to an unsafe design if the user forgets to make those
separate analysis for earthquake or accidental situations.

From now on, the ULS checking is performed for the three situations
at once (uls_permanent, uls_accidental and uls_earthquake) with the
same partial safety factors for materials. If the user wants to make
a separate analysis to make a more accurate calculation (with less
penalizing safety factors for materials) he can make explicit the
design situations to consider for each analysis using the
designSituations parameter of constructor of the limit state class.