Rewrite 3D buckling check routines
There is no buckling in z and y directions, the buckling shape is
determined by the buckling analysis. The new routine compute the
buckling parameters in the direction defined by the eigenvectors
of each element nodes. This approach needs more testing but it is
definitely more robust than relying on axis whose orientation may be
completely unrelated to the buckling phenomenon.