Compute steel shapes auxiliary data out of the module (work in progress)
Computing the auxiliary data (G, Avy, Avz, etc.) each time that
the module is imported is a waste of resources. The scripts that
import the steel shape characteristics from the data provided by
the manufacturer must compute that auxiliary data and write the
data dictionary with it.

All the JSON files older that April 2024 must be updated.