Reorganize soil springs generation on piles
The body of the methods generateSpringsPile*D that generate the spring
elements has been implemented as standalone functions that can be called
outside the Pile classes. That can be useful for problems that for any
reason don't make use of the Pile classes but need to create this kind
of springs.