Move the stiffness retrieval methods to derived classes
The indexes of the different section stiffnesses EA, GAy, GAz
EIz, EIy and GJ in the section stiffness matrix are different
in each class. Instead of guessing its position in the methods
of the PrismaticBarCrossSection class each class has its own
derived methods with the correct indexes hard-coded in them.