Rename tests
Due to the introduction of the getWarpingConstant method, the SteelShape
class has become an abstract class. To avoid calling the abstract
version of the getWarpingConstant method, the old steel shape objects
have been replaced by the correspondent Arcelor steel shapes. As a
consequence, the test_steel_shape have been renamed to match the
code inside.