Test buckling analysis algorithm with more complex models
The eigenvectors of the nodes that are not affected by an
eigenmode are not exactly zero. The previous version of the algorithm
computed directly the normalized versions of those eigenvectors. The
result was a lot of noise that completely masked the true solution
to the problem. This can't be detected with simple structures where
all the nodes participate in every mode.