Reorganize eigenvector normalization routine
The routine has been splitted in two: the first one computes the
maximum infinity norm of the nodes eigenvectors the second one
does the eigenvector normalization only if the computed infinity
norm is different from 1.0 (so the eigenvector have been already
normalized). A third routine normalize the eigenvectors for all
the computed modes.