Defer initialization until the last oportunity
The method initialize_parameters performs the initialization only
if the required values are ready (in particular if the element width
has been assigned and the beam length has not been computed yet).
It is called in setDomain and again in getTangentStiffness. This
way the element properties like element width, gap tolerance, force
tolerance, etc. can be assigned in any moment between the element
creation and its effective use (when the stiffness matrix is computed).