Add information about buckling results
Add the parameter strongAxisBucklingPercent which stores the dot
product of the eigenvector with the element weak axis direction
(normal to the strong axis). A value of 0 for this parameter means
that the buckling bends around the weak axis, a value of 1 means
that the buckling is bending around the strong axis.