Reduce the height of the buckling member and fix section assigment
The mechanical slenderness in the previous version of the test was
outside of the validity range of the method (see clause 43.5 of EHE),
so I have reduced the height of the member to reduce it.

Moreover, I've fixed the assignmentof the RC sections (circular
and flat) to the corresponding members.