Fix problem with small values of bending moments
When the bending moments of the member (without buckling effects) are
very small with alternate negative and positive values along the member
length the addition of the bending moment due to buckling results in a
sequence of not so small positive and negative moments along the member.
The addition of the moments due to buckling amplifies the small preexisting
negative and positive moments which is not the intended effect.

To avoid this problem we check the value of the preexisting moment and
only if it is clearly not zero we add the new moment to it. Otherwise
we replace the small moment with the moment due to buckling directly.