Replace CFncr by alpha_cr
The alpha_cr factor is defined in paragraph (2) of clause 5.2.1 of Eurocode 4
among other codes of practice, so it is a well established concept.
On the other hand, the CFncr factor was a home cooked buckling
related capacity factor (which is exactly the inverse of alpha_cr).
It is preferable to use the name established in the standards, so
I've substituted one factor by the other everywhere in the code.