Create the sections only if needed
The createSections method is modified so it doesn't create the
sections from the given templates unless the internal section
container is empty (the sections are not created yet) or the
forceCreation argument is true. Therefore, the warning message
 concerning already existing creation only appears if there are
already some sections in the internal container and the
forceCreation argument is true.