Use wood armer axial forces by default
Change the default value for woodArmerAlsoForAxialForces to True.
It is safer to set this value to true, and then, if the user wants
to analyze the in-plane shear separately, he can always set this option
to false. As a consequence of this change, two verification tests
have been modified to give the same results they gave prior to this
change.