Expose methods to pick nodes and elements using geometry
The exposed methods are used to select the nodes or elements
inside or outside a 2D or 3D geometric object. The pickElemsCrossing
is used to select the elements that are partially inside a geometric
object. This is useful when computing forces around the critical
perimeter in punching shear strength computations.