Implement Projection and getLengthUpTo methods
Those methods are used to project the positions of the nodes onto
a contour to compute the forces of the elements around the contour
on the contour. The first method is used to compute the projection
of the node position into the perimeter of the contour. The second
method (getLengthUpTo) allows to order the computed projections
along the contour perimeter. This is necessary to compute the
tributary lengths corresponding to each projection and to obtain
a diagram of the forces along the contour perimeter.