Debug subdivision level computation
The previous code was buggy and gave bad results for some cases.
Now the code is much more similar to the procedure exposed in
the article from Robert Schneiders, even if it differs slightly
from it, specially regarding the calculation of the subdivision
level for the edges. If the formulas in page 30 are strictly applied,
I get the wrong subdivision levels for nodes, for example when
reproducing the example in figure 54.