Fix error in getSectionDeformation
The initial section curvature must be substracted from at the element
back end and added to the element front end, otherwise the internal forces
at the element due to those curvatures become zero. The error was
hidden because the curvature in the current set of verification tests
was doubled (the vector argument specifies deformations at points
(0,0), (1,0) and (0,1), so the deformation corresponding to a given
curvature is half the curvature value -so the section deformation
is symmetrical with respect to the center of gravity-).

This problem need to be checked on the other beam elements.