Implement methods to remove constraints from the domain
The previous versions of those methods took the constraint itself
as arguments, so there was no ambiguity with respect to the type of
constraints. The problem with this approach was that if you want
to remove some constraints and put the constraints objects in a
Python list the destructor of the calls the destructor of the
constraints objects and then crashes because the constraints
have already been destroyed. To solve this, we can put the tags
of the constraints to destroy in a list, so the list of integers
can be destroyed smoothly. On the other hand, two constraints of
different type can have the same identifier, so we need a mechanism
to specify no only the constraint tab but also the constraint type.
The solution adopted has been to create parallel methods to remove
the different types of constraints: removeSFreedomConstraint,
removeMFreedomConstraint and removeMRMFreedomConstraint.