Implement Hole class and rewrite section modulus computation
The Hole class allows to define holes inside other sections.
The Wzel and Wyel and Wzel have been rewriten in terms of
moments of inertia (Iz and Iy) and most distant fiber (yMax
and zMax). This approach is much more generic and avoids to
develop a custom recipe for each geometry.