Update code to reflect the last changes
The improvement of the constraint handlers gives much more accurate
results in general but, when there are inactive elements, the error
in the reactions is a bit bigger than before. This is probably due
to the fact that currently, the forces introduced by the "dead"
elements are ignored (but not the stiffnesses to keep the "dead"
nodes "fixed"). Probably the best solution is to reintroduce the
forces of the "dead" elements (small but not zero) again.