Replace pyexcel_ods with openpyxl
The pyexcel_ods library is not available in the Ubuntu apt repositories
and the installation of Python libraries for system-wide use has become
quite cumbersome in Ubuntu 24.04. The adopted solution is to deprecate
the use of pyexcel_ods in favor of openpyxl. This makes necessary to
convert the spreadsheets from Libreoffice (ods format) to Excel
(xlsx format) because openpyxl don't read the former.