Define ElasticPerfectlyPlasticMaterial class
And change the names of setupElasticShear3DSection and
setupElasticShear2DSection by defElasticShearSection3d and
defElasticShearSection2d that are the names used everywhere else
in the code. The old methods have been marked as deprecated but
they still exist.

Concerning the ElasticPerfectlyPlasticMaterial class, it serves
to the same purpose of the BasicElasticMaterial and the other classes
in this module: allows to store the material parameters and defer
the definition of the actual XC material until the FE problem has
been created.