Avoid depth as argument to compute earth pressures
The get_earth_thrusts and get_horizontal_soil_reaction_diagram functions
were written assuming an homogeneous soil when computing the vertical
pressure on the interest point. Both functions have been rewritted on terms
of the vertical pressure, so they can be used more generally. In the case
of an homogeneous soil the vertical pressure can be computed very easily.