Store earth and water pressures instead of forces
The forces on the nodes depend on the tributary area which varies
along the pile wall. As a result, the graphical representation presents
discontinuities that are related only to the element sizes. The
pressure diagrams are much more clear.

Also add designValue as a parameter to the genMesh method to create
the model according to the design values of the soil properties.