Implement clearRCsections method
This method allows to remove from this object the information that
depends on the FE problem object. This allows to remove the pointers
to the objects inside the FEProblem data structuere before freeing
the memory of the FE problem itself (see, for example, the methods
internalForcesVerification3D and internalForcesVerification2D in
RC_material_distribution.py).

Also implement the latexRepor and pdfReport methods.