Avoid solver crash during non-linear analysis of phantom model
When a non-linear analysis of a phantom model is performed the
sections that are not strong enough to withstand the computed internal
forces will make the solver to crash due to a lack of convergence.
To avoid this the capacity factor of the sections is computed and
if its value is greater than 1 (or a user defined value) the section
is not added to the phantom model.