Remove locomotiveLM1 object from EC1_rail_load_models
To have only a "execution-wide" locomotive object is very dangerous.
It its very easy to change one of its parameters unintentionally
(for example reduce its dynamic factor to 0.7 to represent an
accidental load) with disastrous consequences for the values
of the regular (not accidental) loads that get reduced its
value without notice.