Update code
The computed value of Vcu has changed following the fixation of
an error in function getVcuEHE08CrackAngle. As a consequence, the
computed values in this test have been updated. The design value
of the shear force has been incremented to force the optimization
algorithm to search for a bigger shear strength.