Fix error in the computation of lateral earth pressures
The getPressure method of EarthPressureModel was buggy. The earth
lateral pressure coefficient is defined as the s defined as the ratio
of the horizontal effective stress, \sigma'_h, to the vertical effective
stress, \sigma'_v. As a consequence, the algorithm must compute
the vertical effective pressure, and then, multiply by the
lateral earth pressure coefficient.