Make easier elastomeric bearing definition according to EN-1337
Create a JSON file containing the data from  Table 3 of the
standard, and assign a code to each one of the standardized bearings.
The creation of a RectangularLaminatedBearing from the dictionary
is made using the assigned code (just like we do with steel shapes).
The function get_rectangular_bearings_table in the en_1337_bearings
module can be used to obtain a table with all the available bearings.