Fix error in computation of total strain
Instead of adding the values of the strains in the inner and outer
layers (which are different parts of the bearing) the algorithm now
considers the maximum of those values. Concening the strains due
to the shear and the rotation of the bearing the absolute values
of the displacements and rotations have been computed when appropriate.