Add methods returning connected objects that also belong to a set
The methods that return the objects connected to nodes, elements,
points, edges, surfaces and bodies have been versioned to receive
a Set pointer. The new version of the methods return only the connected
objects that also belong to the given set.

The methods have been exposed to Python according to the following
convention:

- The version of the method that returns the connected items regardless
  of wether they belong to a set or not are exposed as properties and the
prefix "get" has been removed.
- The version that receives a set to check if the connected items belong
to it, keeps the "get" prefix and is exposed as a method (as it could not
be otherwise).

The Python modules and verification test need to be updated to reflect
this changes.