Introduce the BeamColumn base class
This class has the getSectionPtr method that returns a pointer to
the element's section material regardless of the element type
(elastic beam column, force beam column, displacement beam
column, etc.). The intention is to provide access to the element
section material from outside the element itself: see, for example,
the getStrainsMatrix method in BeamStrainLoad class.