Ignore incomplete results when displaying limit state result diagrams
Ignore elements that do not have the properties required to display
the diagram. This is useful when for some reason the results have
not been computed in all the elements in the set (for example
bending capacity exceeded when computing shear strength of RC
sections). Instead of letting the graphical routines to crash
the offending elements are ignored and a warning message is
issued.