Implement resetNodalCoordinates method
This method is used to update the element variables that depend on
nodal coordinates after a change in those coordinates (for example,
when introducing a precamber or an initial imperfection in the model).
Tthis method was already implemented for linear elements with the
name initializeCoordTransf (because it computed the new coordinate
transformation matrix with the new nodal coordinates). The new name
is more general and refers to the change of any variable affected
by the nodal coordinates. Some verification examples can be found
in the verif/tests/solution/initial_imperfection folder.